agents/runner/runtime/pyproject.toml
[project]
name = "flowkit-runtime"
version = "0.0.1"
description = "FlowKit runtime implementations (LangGraph flows)"
requires-python = ">=3.14,<3.15"
dependencies = [
  "pydantic>=2.9",
  "langgraph>=0.2.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "mypy>=1.12",
  "ruff>=0.6",
  "black>=24.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
warn_return_any = true

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

